fix: move all scooter stuff to a sub compoinent
